Extracted class to their own files Added EntityCreated log message
