+++ ECS/ECS.Core/DL.ECS.Core.csproj
@@ -43,9 +43,11 @@
     <Reference Include="System.Xml" />
   </ItemGroup>
   <ItemGroup>
+    <Compile Include="Components\ComponentId.cs" />
     <Compile Include="Components\ComponentManager.cs" />
-    <Compile Include="Context.cs" />
-    <Compile Include="Entity.cs" />
+    <Compile Include="Context\Context.cs" />
+    <Compile Include="Entity\Entity.cs" />
+    <Compile Include="Entity\EntityId.cs" />
     <Compile Include="Exceptions\ECSException.cs" />
     <Compile Include="Exceptions\Entity\EntityAlreadyHasComponentException.cs" />
     <Compile Include="Exceptions\Entity\EntityDoesNotHaveComponentException.cs" />
@@ -55,8 +57,10 @@
     <Compile Include="Components\IComponent.cs" />
     <Compile Include="Properties\AssemblyInfo.cs" />
     <Compile Include="Relations\IRelation.cs" />
+    <Compile Include="Relations\RelationId.cs" />
     <Compile Include="Relations\RelationManager.cs" />
     <Compile Include="Relations\Set.cs" />
+    <Compile Include="SystemNotification.cs" />
     <Compile Include="Systems\ISystem.cs" />
   </ItemGroup>
   <ItemGroup>
