+++ ECS/ECS.Core/SystemNotification.cs
@@ -1,0 +1,11 @@
+﻿using DL.Infrastructure;
+
+namespace DL.ECS.Core
+{
+    public static class SystemNotificationExtensions
+    {
+        public static void EntityCreated(this ISystemNotification systemNotification,
+            IEntity entity)
+            => systemNotification.Logger.Information($"Entity created: {entity}");
+    }
+}
