+++ Infrastructure/DL.Infrastructure/packages.config
@@ -1,4 +1,5 @@
 ﻿<?xml version="1.0" encoding="utf-8"?>
 <packages>
   <package id="Serilog" version="2.4.0" targetFramework="net461" />
+  <package id="Serilog.Sinks.Trace" version="2.1.0" targetFramework="net461" />
 </packages>
